refactor(storage): Remove unused code
