+++ pkg/storage/vulnerability.go
@@ -1,403 +1,834 @@
 package storage
 
 import (
+	"context"
 	"database/sql"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"time"
+	"log/slog"
+	"strings"
 
+	"github.com/cockroachdb/apd/v3"
 	"github.com/google/uuid"
+	"github.com/kptm-tools/common/common/pkg/enums"
 	"github.com/kptm-tools/common/common/pkg/results/tools"
 	"github.com/kptm-tools/core-service/pkg/customerrors"
 	"github.com/kptm-tools/core-service/pkg/domain"
+	"github.com/kptm-tools/core-service/pkg/interfaces"
+	"github.com/kptm-tools/core-service/pkg/repository"
+	"github.com/sqlc-dev/pqtype"
 )
 
-func (s *PostgreSQLStore) CreateScanVulnerability(
-	tx *sql.Tx,
+type VulnerRepo struct {
+	db             *sql.DB
+	defaultQueries *repository.Queries
+}
+
+var _ interfaces.VulnerabilityRepository = (*VulnerRepo)(nil)
+
+func NewVulnerRepository(queries *repository.Queries, db *sql.DB) *VulnerRepo {
+	return &VulnerRepo{
+		defaultQueries: queries,
+		db:             db,
+	}
+}
+
+// getQueries retrieves the correct *repository.Queries instance from the context.
+// If a transaction is active, it gets the transactional queries. Otherwise, it uses
+// the defaultQueries.
+func (r *VulnerRepo) getQueries(ctx context.Context) *repository.Queries {
+	return GetQueriesFromContext(ctx, r.defaultQueries)
+}
+
+func (r *VulnerRepo) CreateVulnerability(
+	ctx context.Context,
+	hostID uuid.UUID,
 	scanID uuid.UUID,
-	hostID int,
-	serviceID *int,
-	operatingSystemID *int,
 	vuln tools.Vulnerability,
-) error {
-	var execer interface {
-		QueryRow(query string, args ...any) *sql.Row
-		Exec(query string, args ...any) (sql.Result, error)
+	vulnType repository.VulnerabilityTypeEnum,
+) (*tools.Vulnerability, error) {
+	query := r.getQueries(ctx)
+
+	params := repository.CreateVulnerabilityParams{
+		HostID:         hostID,
+		ScanID:         scanID,
+		CveID:          sql.NullString{String: vuln.CveID, Valid: vuln.CveID != ""},
+		Title:          vuln.CveID,
+		Description:    sql.NullString{String: vuln.Description, Valid: vuln.Description != ""},
+		Severity:       vuln.BaseSeverity.String(),
+		VulnSource:     "NVD",
+		VulnType:       vulnType,
+		AnalystComment: sql.NullString{String: "", Valid: false},
 	}
-
-	if tx != nil {
-		execer = tx
-	} else {
-		execer = s.db
-	}
-
-	// 1. Insert vulnerability to vulners table
-	query := `
-    INSERT INTO scan_vulnerabilities (
-            scan_id,
-            host_id,
-            service_id, 
-            operating_system_id,
-            type,
-            vulnerability_id,
-            cvss,
-            vuln_references,
-            exploit_score,
-            exploitability,
-            description,
-            access_type,
-            complexity,
-            privileges_required,
-            likelihood,
-            risk_score,
-            impact_score,
-            integrity_impact,
-            availability_impact,
-            base_severity,
-            vendor_comments,
-            published,
-            last_updated
-  ) VALUES (
-            $1,  -- scan_id
-            $2,  -- host_id
-            $3,  -- service_id
-            $4,  -- operating_system_id
-            $5,  -- type
-            $6,  -- vulnerability_id
-            $7,  -- cvss
-            $8,  -- vuln_references
-            $9,  -- exploit_score 
-            $10, -- exploitability
-            $11, -- description
-            $12, -- access_type
-            $13, -- complexity
-            $14, -- privileges_required
-            $15, -- likelihood
-            $16, -- risk_score
-            $17, -- impact_score
-            $18, -- integrity_impact
-            $19, -- availability_impact
-            $20, -- base_severity 
-            $21,  -- vendor_comments
-            $22, -- published
-            $23  -- last_updated
-  )
-  `
-
-	vulnReferencesJSON, err := json.Marshal(vuln.References)
+	dbVulner, err := query.CreateVulnerability(ctx, params)
 	if err != nil {
-		return fmt.Errorf("failed to marshal vuln references to JSON: %w", err)
+		return nil, err
 	}
-
-	vendorCommentsJSON, err := json.Marshal(vuln.VendorComments)
-	if err != nil {
-		return fmt.Errorf("failed to marshal vendor comments to JSON: %w", err)
-	}
-
-	_, err = execer.Exec(
-		query,
-		scanID,
-		hostID,
-		serviceID,
-		operatingSystemID,
-		vuln.Type.String(),
-		vuln.ID,
-		vuln.BaseCVSSScore,
-		vulnReferencesJSON,
-		vuln.Exploit.Score,
-		vuln.Exploit.Exploitability,
-		vuln.Description,
-		vuln.Access.String(),
-		vuln.Complexity.String(),
-		vuln.PrivilegesRequired.String(),
-		vuln.Likelihood.String(),
-		vuln.RiskScore,
-		vuln.ImpactScore,
-		vuln.IntegrityImpact.String(),
-		vuln.AvailabilityImpact.String(),
-		vuln.BaseSeverity.String(),
-		vendorCommentsJSON,
-		vuln.Published,
-		vuln.LastUpdated,
-	)
-	if err != nil {
-		return fmt.Errorf("failed to exec insertion: %w", err)
-	}
-
-	return nil
+	domVuln := toDomainVuln(dbVulner)
+	return &domVuln, err
 }
 
-func (s *PostgreSQLStore) CreateOSVulnerability(
-	tx *sql.Tx,
-	scanID uuid.UUID,
-	hostID int,
-	operatingSystemID int,
-	vuln tools.Vulnerability,
-) error {
-	osIDPtr := &operatingSystemID
-	var serviceIDPtr *int = nil
-
-	return s.CreateScanVulnerability(tx, scanID, hostID, serviceIDPtr, osIDPtr, vuln)
-}
-
-func (s *PostgreSQLStore) CreateServiceVulnerability(
-	tx *sql.Tx,
-	scanID uuid.UUID,
-	hostID int,
-	serviceID int,
-	vuln tools.Vulnerability,
-) error {
-	serviceIDPtr := &serviceID
-	var osIDPtr *int = nil
-
-	return s.CreateScanVulnerability(tx, scanID, hostID, serviceIDPtr, osIDPtr, vuln)
-}
-
-func (s *PostgreSQLStore) GetScanVulnerabilities(scanID uuid.UUID) ([]*domain.Vulnerability, error) {
-	query := `
-      SELECT 
-          id,
-          vulnerability_id,
-          scan_id,
-          host_id,
-          "type",
-          cvss,
-          vuln_references,
-          created_at,
-          updated_at,
-          severity,
-          analyst_comment,
-          vendor_comments,
-          operating_system_id,
-          service_id,
-          description,
-          access_type,
-          complexity,
-          privileges_required,
-          likelihood,
-          risk_score,
-          impact_score,
-          exploit_score,
-          exploitability,
-          integrity_impact,
-          availability_impact,
-          base_severity,
-          published,
-          last_updated
-      FROM scan_vulnerabilities
-      WHERE scan_id = $1
-  `
-
-	rows, err := s.db.Query(query, scanID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to query rows: %w", err)
-	}
-	vulners := make([]*domain.Vulnerability, 0)
-	for rows.Next() {
-		var vuln domain.Vulnerability
-		var referencesBytes []byte
-		var vendorCommentsBytes []byte
-		var exploit tools.Exploit
-
-		err := rows.Scan(
-			&vuln.ID,
-			&vuln.VulnerabilityID,
-			&vuln.ScanID,
-			&vuln.HostID,
-			&vuln.Type,
-			&vuln.BaseCVSSScore,
-			&referencesBytes,
-			&vuln.CreatedAt,
-			&vuln.UpdatedAt,
-			&vuln.Severity,
-			&vuln.AnalystComment,
-			&vendorCommentsBytes,
-			&vuln.OperatingSystemID,
-			&vuln.ServiceID,
-			&vuln.Description,
-			&vuln.AccessType,
-			&vuln.Complexity,
-			&vuln.PrivilegesRequired,
-			&vuln.Likelihood,
-			&vuln.RiskScore,
-			&vuln.ImpactScore,
-			&exploit.Score,
-			&exploit.Exploitability,
-			&vuln.IntegrityImpact,
-			&vuln.AvailabilityImpact,
-			&vuln.BaseSeverity,
-			&vuln.Published,
-			&vuln.LastUpdated,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan into domain vulnerability: %w", err)
-		}
-
-		// Unmarshal and assign references
-		if err := json.Unmarshal(referencesBytes, &vuln.References); err != nil {
-			return nil, fmt.Errorf("failed to unmarshal vuln_references: %w", err)
-		}
-
-		// Unmarshal vendor comments
-		if err := json.Unmarshal(vendorCommentsBytes, &vuln.VendorComments); err != nil {
-			return nil, fmt.Errorf("failed to unmarshal vendor_comments: %w", err)
-		}
-
-		vuln.Exploit = exploit
-
-		vulners = append(vulners, &vuln)
-	}
-	if rows.Err() != nil {
-		return nil, fmt.Errorf("encountered error during row iteration: %w", err)
-	}
-
-	return vulners, nil
-}
-
-func (s *PostgreSQLStore) GetVulnerabilityByID(vulnID int) (*domain.Vulnerability, error) {
-	query := `
-      SELECT 
-          id,
-          vulnerability_id,
-          scan_id,
-          host_id,
-          "type",
-          cvss,
-          vuln_references,
-          created_at,
-          updated_at,
-          severity,
-          analyst_comment,
-          vendor_comments,
-          operating_system_id,
-          service_id,
-          description,
-          access_type,
-          complexity,
-          privileges_required,
-          likelihood,
-          risk_score,
-          impact_score,
-          exploit_score,
-          exploitability,
-          integrity_impact,
-          availability_impact,
-          base_severity,
-          published,
-          last_updated
-      FROM scan_vulnerabilities
-      WHERE id = $1
-  `
-
-	var vuln domain.Vulnerability
-	var referencesBytes []byte
-	var vendorCommentsBytes []byte
-	var exploit tools.Exploit
-
-	err := s.db.QueryRow(query, vulnID).Scan(
-		&vuln.ID,
-		&vuln.VulnerabilityID,
-		&vuln.ScanID,
-		&vuln.HostID,
-		&vuln.Type,
-		&vuln.BaseCVSSScore,
-		&referencesBytes,
-		&vuln.CreatedAt,
-		&vuln.UpdatedAt,
-		&vuln.Severity,
-		&vuln.AnalystComment,
-		&vendorCommentsBytes,
-		&vuln.OperatingSystemID,
-		&vuln.ServiceID,
-		&vuln.Description,
-		&vuln.AccessType,
-		&vuln.Complexity,
-		&vuln.PrivilegesRequired,
-		&vuln.Likelihood,
-		&vuln.RiskScore,
-		&vuln.ImpactScore,
-		&exploit.Score,
-		&exploit.Exploitability,
-		&vuln.IntegrityImpact,
-		&vuln.AvailabilityImpact,
-		&vuln.BaseSeverity,
-		&vuln.Published,
-		&vuln.LastUpdated,
-	)
+func (r *VulnerRepo) GetVulnerabilityByID(
+	ctx context.Context,
+	vulnID uuid.UUID,
+) (*tools.Vulnerability, error) {
+	queries := r.getQueries(ctx)
+	dbVulner, err := queries.GetVulnerabilityWithCveDetailByID(ctx, vulnID)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
-			return nil, fmt.Errorf("%w %d: %w", customerrors.ErrVulnNotFound, vulnID, err)
+			return nil, customerrors.ErrVulnNotFound
 		}
-		return nil, fmt.Errorf("error scanning into vulnerability: %w", err)
+		return nil, fmt.Errorf("failed to query for vulnerability", err)
 	}
 
-	if err := json.Unmarshal(referencesBytes, &vuln.References); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal referencesBytes: %w", err)
-	}
-
-	if err := json.Unmarshal(vendorCommentsBytes, &vuln.VendorComments); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal vendorCommentsBytes: %w", err)
-	}
-
-	vuln.Exploit = exploit
-
-	return &vuln, nil
+	domVuln := toDomainVulnerabilityWithCveDetail(dbVulner)
+	return &domVuln, nil
 }
 
-func (s *PostgreSQLStore) GetScanVulnerabilityCount(scanID uuid.UUID) (int, error) {
-	query := `
-    SELECT COUNT(id)
-    FROM scan_vulnerabilities
-    WHERE scan_id = $1
-  `
-
-	var vulnCount int
-	err := s.db.QueryRow(query, scanID).Scan(&vulnCount)
+func (r *VulnerRepo) GetVulnerabilitiesWithCveDetailByScanID(ctx context.Context, scanID uuid.UUID) (
+	[]tools.Vulnerability,
+	error,
+) {
+	queries := r.getQueries(ctx)
+	cveDetailedVulns, err := queries.GetVulnerabilitiesWithCveDetailsByScanID(ctx, scanID)
 	if err != nil {
-		return 0, fmt.Errorf("failed to query row: %w", err)
+		return nil, err
 	}
 
-	return vulnCount, nil
+	domainVulns := make([]tools.Vulnerability, len(cveDetailedVulns))
+	for i, cveDetailedVuln := range cveDetailedVulns {
+		domainVulns[i] = toDomainVulnWithCveDetail(cveDetailedVuln)
+	}
+	return domainVulns, nil
 }
 
-func (s *PostgreSQLStore) HasComment(vulnID int) (bool, error) {
-	var exists bool
-	query := `SELECT CASE WHEN analyst_comment is not null  THEN true ELSE false END FROM scan_vulnerabilities
-             WHERE id=$1`
-	err := s.db.QueryRow(query, vulnID).Scan(&exists)
+func (r *VulnerRepo) UpdateVulnerabilityComment(ctx context.Context, vulnID uuid.UUID, newComment string) (bool, error) {
+	queries := r.getQueries(ctx)
+	params := repository.UpdateVulnerabilityCommentParams{
+		ID:             vulnID,
+		AnalystComment: sql.NullString{String: newComment, Valid: true},
+	}
+	_, err := queries.UpdateVulnerabilityComment(ctx, params)
 	if err != nil {
-		return false, fmt.Errorf("failed to query existence: %w", err)
-	}
-
-	return exists, nil
-}
-
-func (s *PostgreSQLStore) UpdateVulnerabilityComment(vulnID int, comment string) (bool, error) {
-	query := `
-    UPDATE scan_vulnerabilities SET
-         analyst_comment =$2 , updated_at=$3
-             WHERE id= $1`
-	res, err := s.db.Exec(query, vulnID, comment, time.Now().UTC())
-	switch err {
-	case nil:
-		count, _ := res.RowsAffected()
-		return count == 1, nil
-	default:
 		return false, err
 	}
+	return true, nil
 }
 
-func (s *PostgreSQLStore) DeleteVulnerabilityComment(vulnID int) (bool, error) {
-	query := `UPDATE scan_vulnerabilities SET analyst_comment= NULL, updated_at=$2 
-             WHERE id= $1`
-
-	res, err := s.db.Exec(query, vulnID, time.Now().UTC())
-	switch err {
-	case nil:
-		count, _ := res.RowsAffected()
-		return count == 1, nil
-	default:
+func (r *VulnerRepo) DeleteVulnerabilityComment(ctx context.Context, vulnID uuid.UUID) (bool, error) {
+	queries := r.getQueries(ctx)
+	_, err := queries.DeleteVulnerabilityComment(ctx, vulnID)
+	if err != nil {
 		return false, err
 	}
+	return true, nil
+}
+
+func (r *VulnerRepo) HasComment(ctx context.Context, vulnID uuid.UUID) (bool, error) {
+	queries := r.getQueries(ctx)
+	vulnerStatus, err := queries.CheckVulnerabilityHasComment(ctx, vulnID)
+	if err != nil {
+		return false, err
+	}
+
+	hasComment, ok := vulnerStatus.HasComment.(bool)
+	if !ok {
+		return false, fmt.Errorf(
+			"expected 'HasComment' to be of type bool, but got %T (value: %v)",
+			vulnerStatus.HasComment,
+			vulnerStatus.HasComment,
+		)
+	}
+	return hasComment, nil
+}
+
+func (r *VulnerRepo) GetVulnerabilityType(ctx context.Context, vulnID uuid.UUID) (repository.VulnerabilityTypeEnum, error) {
+	queries := r.getQueries(ctx)
+	vuln, err := queries.GetVulnerabilityByID(ctx, vulnID)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return repository.VulnerabilityTypeEnumNETWORKOS, customerrors.ErrVulnNotFound
+		}
+		return repository.VulnerabilityTypeEnumNETWORKOS, err
+	}
+	return vuln.VulnType, nil
+}
+
+func (r *VulnerRepo) CreateNetworkOSVulnerabilityForOS(
+	ctx context.Context,
+	vulnID, scanID, hostID uuid.UUID,
+	operatingSystemID int32,
+) error {
+	queries := r.getQueries(ctx)
+	params := repository.CreateNetworkOSVulnerabilityParams{
+		VulnerabilityID:   vulnID,
+		ScanID:            scanID,
+		HostID:            hostID,
+		OperatingSystemID: sql.NullInt32{Int32: operatingSystemID, Valid: true},
+		ServiceID:         sql.NullInt32{Valid: false},
+	}
+	_, err := queries.CreateNetworkOSVulnerability(ctx, params)
+	return err
+}
+
+func (r *VulnerRepo) CreateNetworkOSVulnerabilityForService(
+	ctx context.Context,
+	vulnID, scanID, hostID uuid.UUID,
+	serviceID int32,
+) error {
+	queries := r.getQueries(ctx)
+	params := repository.CreateNetworkOSVulnerabilityParams{
+		VulnerabilityID:   vulnID,
+		ScanID:            scanID,
+		HostID:            hostID,
+		OperatingSystemID: sql.NullInt32{Valid: false},
+		ServiceID:         sql.NullInt32{Int32: serviceID, Valid: true},
+	}
+	_, err := queries.CreateNetworkOSVulnerability(ctx, params)
+	return err
+}
+
+func (r *VulnerRepo) GetNetworkOSVulnerability(ctx context.Context, vulnID uuid.UUID) (*domain.NetworkOSVulnerability, error) {
+	queries := r.getQueries(ctx)
+	dbNetOSVuln, err := queries.GetNetworkOSVulnerabilityByID(ctx, vulnID)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return nil, customerrors.ErrVulnNotFound
+		}
+	}
+
+	// Handle nil values
+	var osID, svcID *int32
+	if dbNetOSVuln.OperatingSystemID.Valid {
+		osID = &dbNetOSVuln.OperatingSystemID.Int32
+	}
+	if dbNetOSVuln.ServiceID.Valid {
+		svcID = &dbNetOSVuln.ServiceID.Int32
+	}
+
+	domNetOSVuln := domain.NetworkOSVulnerability{
+		VulnerabilityID:   dbNetOSVuln.VulnerabilityID,
+		ScanID:            dbNetOSVuln.ScanID,
+		HostID:            dbNetOSVuln.HostID,
+		OperatingSystemID: osID,
+		ServiceID:         svcID,
+		CreatedAt:         &dbNetOSVuln.CreatedAt.Time,
+		UpdatedAt:         &dbNetOSVuln.UpdatedAt.Time,
+	}
+	return &domNetOSVuln, nil
+}
+
+func (r *VulnerRepo) GetSeverityCountsByScanID(ctx context.Context, scanID uuid.UUID) (tools.SeverityCounts, error) {
+	queries := r.getQueries(ctx)
+
+	dbCounts, err := queries.GetVulnerabilitySeverityCountsByScanID(ctx, scanID)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return tools.SeverityCounts{}, customerrors.ErrScanNotFound
+		}
+		return tools.SeverityCounts{}, err
+	}
+
+	counts := tools.SeverityCounts{
+		Unknown:  int(dbCounts.UnknownVulnerabilities),
+		None:     int(dbCounts.NoneVulnerabilities),
+		Low:      int(dbCounts.LowVulnerabilities),
+		Medium:   int(dbCounts.MediumVulnerabilities),
+		High:     int(dbCounts.HighVulnerabilities),
+		Critical: int(dbCounts.CriticalVulnerabilities),
+	}
+	return counts, nil
+}
+
+func (r *VulnerRepo) GetScanVulnerabilityAggregates(
+	ctx context.Context,
+	params domain.VulnerabilityAggregatesParams,
+) (*domain.VulnerabilityAggregatesResult, error) {
+	queries := r.getQueries(ctx)
+	sqlcParams := repository.GetScanVulnerabilityAggregatesByScanIDParams{
+		ScanID: params.ScanID,
+	}
+	if len(params.SeverityFilters) > 0 {
+		sqlcParams.SeverityFilters = params.SeverityFilters
+	} else {
+		sqlcParams.SeverityFilters = nil
+	}
+	row, err := queries.GetScanVulnerabilityAggregatesByScanID(ctx, sqlcParams)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return nil, customerrors.ErrScanNotFound
+		}
+		return nil, err
+	}
+	return &domain.VulnerabilityAggregatesResult{
+		TotalVulnerabilities: int(row.TotalVulnerabilities),
+		SeverityCounts: tools.SeverityCounts{
+			Critical: int(row.CriticalVulnerabilities), High: int(row.HighVulnerabilities),
+			Medium: int(row.MediumVulnerabilities), Low: int(row.MediumVulnerabilities),
+			None: int(row.NoneVulnerabilities), Unknown: int(row.UnknownVulnerabilities),
+		},
+	}, nil
+}
+
+func (r *VulnerRepo) GetVulnerabilityCategoriesByScan(
+	ctx context.Context,
+	params domain.VulnerabilityCategoriesParams,
+) ([]domain.ServiceCategoryData, error) {
+	queries := r.getQueries(ctx)
+	sqlcParams := repository.GetVulnerabilityCategoriesByScanParams{
+		ScanID: params.ScanID,
+	}
+	if len(params.SeverityFilters) > 0 {
+		sqlcParams.SeverityFilters = params.SeverityFilters
+	} else {
+		sqlcParams.SeverityFilters = nil
+	}
+
+	rows, err := queries.GetVulnerabilityCategoriesByScan(ctx, sqlcParams)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return []domain.ServiceCategoryData{}, nil
+		}
+		return nil, err
+	}
+	categories := make([]domain.ServiceCategoryData, len(rows))
+	for i, row := range rows {
+		categories[i] = domain.ServiceCategoryData{
+			Category: row.Category.String,
+			Count:    int(row.Count),
+		}
+	}
+	return categories, nil
+}
+
+func (r *VulnerRepo) GetVulnerabilityCountByScanID(ctx context.Context, scanID uuid.UUID) (int, error) {
+	queries := r.getQueries(ctx)
+	count, err := queries.GetVulnerabilityCountByScanID(ctx, scanID)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return 0, customerrors.ErrScanNotFound
+		}
+		return 0, err
+	}
+	return int(count), nil
+}
+
+func (r *VulnerRepo) GetHostVulnerabilityTrends(
+	ctx context.Context,
+	params domain.VulnerabilityTrendsParams,
+) ([]domain.ServiceTimePeriod, error) {
+	// This query is kind of complicated, but what it does is fill out time_periods and labels,
+	// even when there is no data for said time_period. E.g: we only have data
+	// for February, but we want to have the counts for other months to be 0 too.
+	// It also has a CTE called ScanPeriods, which identifies if there was a scan at all
+	// during that period. This allows us to differentiate if we got a 0 count because there
+	// are no vulnerabilities, or because there is no scan.
+	baseTrendQuery := `
+		WITH TimePeriods AS (
+			SELECT
+				CASE
+					WHEN $2 = 'Month' THEN TO_CHAR(date_series, 'FMMonth')
+					WHEN $2 = 'Quarter' THEN 'Q' || TO_CHAR(date_series, 'Q')
+					WHEN $2 = 'Semester' THEN 'Semester ' || CASE WHEN TO_CHAR(date_series, 'MM')::integer <= 6 THEN '1' ELSE '2' END
+					ELSE 'Unknown Period'
+				END AS time_period_label,
+				CASE
+					WHEN $2 = 'Month' THEN TO_CHAR(date_series, 'YYYY-MM')
+					WHEN $2 = 'Quarter' THEN TO_CHAR(date_series, 'YYYY-Q')
+					WHEN $2 = 'Semester' THEN CASE WHEN TO_CHAR(date_series, 'MM')::integer <= 6 THEN '1' ELSE '2' END
+					ELSE '1'
+				END AS ordering_period,
+				date_series
+			FROM generate_series(
+				DATE_TRUNC('year', CURRENT_DATE),
+				DATE_TRUNC('year', CURRENT_DATE) + INTERVAL '1 year' - INTERVAL '1 day',
+				CASE
+					WHEN $2 = 'Month' THEN INTERVAL '1 month'
+					WHEN $2 = 'Quarter' THEN INTERVAL '3 month'
+					WHEN $2 = 'Semester' THEN INTERVAL '6 month'
+					ELSE INTERVAL '1 month'
+				END
+			) AS date_series
+		),
+		ScanPeriods AS (
+			SELECT
+        DISTINCT ON (time_period_label)
+				CASE
+					WHEN $2 = 'Month' THEN TO_CHAR(s.started_at, 'FMMonth')
+					WHEN $2 = 'Quarter' THEN 'Q' || TO_CHAR(s.started_at, 'Q')
+					WHEN $2 = 'Semester' THEN 'Semester ' || CASE WHEN TO_CHAR(s.started_at, 'MM')::integer <= 6 THEN '1' ELSE '2' END
+					ELSE 'Unknown Period'
+				END AS time_period_label,
+				s.id AS scan_id,
+				s.started_at AS scan_started_at
+			FROM scans s
+			WHERE s.host_id = $1
+				AND EXTRACT(YEAR FROM s.started_at) = EXTRACT(YEAR FROM CURRENT_DATE)
+      ORDER BY time_period_label, s.started_at DESC
+		),
+		VulnerabilityCounts AS (
+			SELECT
+				sp.time_period_label,
+				COUNT(sv.id) AS vulnerability_count
+			FROM ScanPeriods sp
+			LEFT JOIN vulnerabilities sv ON sv.scan_id = sp.scan_id
+			INNER JOIN scans s ON sp.scan_id = s.id
+			WHERE s.host_id = $1
+				AND EXTRACT(YEAR FROM s.started_at) = EXTRACT(YEAR FROM CURRENT_DATE)
+				-- Severity Filter Dynamic Condition goes here
+				%s
+			GROUP BY sp.time_period_label
+		)
+		SELECT
+			tp.time_period_label AS time_period,
+			vc.vulnerability_count AS vulnerability_count -- Now vc.vulnerability_count will be NULL if no scan
+		FROM TimePeriods tp
+		LEFT JOIN ScanPeriods sp ON tp.time_period_label = sp.time_period_label -- Join with ScanPeriods to ensure time period has a scan
+		LEFT JOIN VulnerabilityCounts vc ON tp.time_period_label = vc.time_period_label
+		ORDER BY tp.ordering_period;
+	`
+
+	trendQueryParams := []any{params.HostID, params.TimePeriodFilter.String()}
+	trendSeverityWhereClause, trendQueryParams := buildSeverityWhereClause(params.SeverityFilters, trendQueryParams)
+	forattedTrendQuery := fmt.Sprintf(baseTrendQuery, trendSeverityWhereClause)
+	slog.Debug("Executing Host Trend Query",
+		slog.Any("query_params", trendQueryParams))
+
+	trendsRows, err := r.db.Query(forattedTrendQuery, trendQueryParams...)
+	if err != nil {
+		return nil, fmt.Errorf("failed to query vulnerability trends for host %s: %w", params.HostID.String(), err)
+	}
+	defer trendsRows.Close()
+
+	var timePeriods []domain.ServiceTimePeriod
+	for trendsRows.Next() {
+		var timePeriodData domain.ServiceTimePeriod
+		if err := trendsRows.Scan(&timePeriodData.TimePeriod, &timePeriodData.VulnerabilityCount); err != nil {
+			return nil, fmt.Errorf("failed to scan into time period: %w", err)
+		}
+		timePeriods = append(timePeriods, timePeriodData)
+	}
+	if trendsRows.Err() != nil {
+		return nil, fmt.Errorf("failed to iterate trend rows: %w", err)
+	}
+	return timePeriods, nil
+}
+
+func toDomainVuln(dbVuln repository.Vulnerability) tools.Vulnerability {
+	return tools.Vulnerability{
+		ID:     dbVuln.ID,
+		ScanID: dbVuln.ScanID,
+		CveID:  dbVuln.CveID.String,
+	}
+}
+
+// toDomainVulnerabilityWithCveDetail converts a sqlc-generated GetVulnerabilityWithCveDetailByIDRow
+// into a domain.Vulnerability.
+// It handles null values and type conversions.
+func toDomainVulnerabilityWithCveDetail(dbRow repository.GetVulnerabilityWithCveDetailByIDRow) tools.Vulnerability {
+	// Initialize with direct assignments and default values
+	domVuln := tools.Vulnerability{
+		ID:             dbRow.ID,
+		HostID:         dbRow.HostID,
+		ScanID:         dbRow.ScanID,
+		CveID:          nullStringToString(dbRow.CveID),
+		Description:    nullStringToString(dbRow.Description),
+		AnalystComment: nullStringToString(dbRow.AnalystComment),
+		Published:      nullTimeToTime(dbRow.PublishedDate),
+		LastUpdated:    nullTimeToTime(dbRow.LastModifiedDate),
+		Metrics:        []tools.CVSSMetric{},
+	}
+
+	intermediateData := IntermediateCVEDetailData{
+		DetailID:         dbRow.ID_2,    // c.id
+		DetailCveID:      dbRow.CveID_2, // c.cve_id
+		Cwe:              dbRow.Cwe,
+		PublishedDate:    dbRow.PublishedDate,
+		LastModifiedDate: dbRow.LastModifiedDate,
+		NvdDescription:   dbRow.NvdDescription,
+
+		CvssV2Vector:                dbRow.CvssV2Vector,
+		CvssV2BaseScore:             dbRow.CvssV2BaseScore,
+		CvssV2BaseSeverity:          dbRow.CvssV2BaseSeverity,
+		CvssV2ExploitabilityScore:   dbRow.CvssV2ExploitabilityScore,
+		CvssV2ImpactScore:           dbRow.CvssV2ImpactScore,
+		CvssV2AccessVector:          dbRow.CvssV2AccessVector,
+		CvssV2AccessComplexity:      dbRow.CvssV2AccessComplexity,
+		CvssV2Authentication:        dbRow.CvssV2Authentication,
+		CvssV2ConfidentialityImpact: dbRow.CvssV2ConfidentialityImpact,
+		CvssV2IntegrityImpact:       dbRow.CvssV2IntegrityImpact,
+		CvssV2AvailabilityImpact:    dbRow.CvssV2AvailabilityImpact,
+
+		CvssV30Vector:                dbRow.CvssV30Vector,
+		CvssV30BaseScore:             dbRow.CvssV30BaseScore,
+		CvssV30BaseSeverity:          dbRow.CvssV30BaseSeverity,
+		CvssV30ExploitabilityScore:   dbRow.CvssV30ExploitabilityScore,
+		CvssV30ImpactScore:           dbRow.CvssV30ImpactScore,
+		CvssV30AttackVector:          dbRow.CvssV30AttackVector,
+		CvssV30AttackComplexity:      dbRow.CvssV30AttackComplexity,
+		CvssV30PrivilegesRequired:    dbRow.CvssV30PrivilegesRequired,
+		CvssV30UserInteraction:       dbRow.CvssV30UserInteraction,
+		CvssV30Scope:                 dbRow.CvssV30Scope,
+		CvssV30ConfidentialityImpact: dbRow.CvssV30ConfidentialityImpact,
+		CvssV30IntegrityImpact:       dbRow.CvssV30IntegrityImpact,
+		CvssV30AvailabilityImpact:    dbRow.CvssV30AvailabilityImpact,
+
+		CvssV31Vector:                dbRow.CvssV31Vector,
+		CvssV31BaseScore:             dbRow.CvssV31BaseScore,
+		CvssV31BaseSeverity:          dbRow.CvssV31BaseSeverity,
+		CvssV31ExploitabilityScore:   dbRow.CvssV31ExploitabilityScore,
+		CvssV31ExploitCodeMaturity:   dbRow.CvssV31ExploitCodeMaturity,
+		CvssV31ImpactScore:           dbRow.CvssV31ImpactScore,
+		CvssV31AttackVector:          dbRow.CvssV31AttackVector,
+		CvssV31AttackComplexity:      dbRow.CvssV31AttackComplexity,
+		CvssV31PrivilegesRequired:    dbRow.CvssV31PrivilegesRequired,
+		CvssV31UserInteraction:       dbRow.CvssV31UserInteraction,
+		CvssV31Scope:                 dbRow.CvssV31Scope,
+		CvssV31ConfidentialityImpact: dbRow.CvssV31ConfidentialityImpact,
+		CvssV31IntegrityImpact:       dbRow.CvssV31IntegrityImpact,
+		CvssV31AvailabilityImpact:    dbRow.CvssV31AvailabilityImpact,
+
+		EpssScore:      dbRow.EpssScore,
+		EpssPercentile: dbRow.EpssPercentile,
+		RiskScore:      dbRow.RiskScore,
+		Likelihood:     dbRow.Likelihood,
+
+		EvaluatorComment:  dbRow.EvaluatorComment,
+		EvaluatorImpact:   dbRow.EvaluatorImpact,
+		EvaluatorSolution: dbRow.EvaluatorSolution,
+
+		NvdReferences:  dbRow.NvdReferences,
+		VendorComments: dbRow.VendorComments,
+	}
+
+	if intermediateData.DetailID.Valid {
+		populateVulnerabilityFromIntermediateDetails(&domVuln, &intermediateData)
+	} else {
+		// In case cve_details does not exists due to LEFT JOIN or some shenanigans
+		domVuln.BaseSeverity = enums.SeverityTypeUnknown
+	}
+	return domVuln
+}
+
+func toDomainVulnWithCveDetail(dbRow repository.GetVulnerabilitiesWithCveDetailsByScanIDRow) tools.Vulnerability {
+	// Initialize with direct assignments and default values
+	domVuln := tools.Vulnerability{
+		ID:             dbRow.ID,
+		HostID:         dbRow.HostID,
+		ScanID:         dbRow.ScanID,
+		CveID:          nullStringToString(dbRow.CveID),
+		Description:    nullStringToString(dbRow.Description),
+		AnalystComment: nullStringToString(dbRow.AnalystComment),
+		Published:      nullTimeToTime(dbRow.PublishedDate),
+		LastUpdated:    nullTimeToTime(dbRow.LastModifiedDate),
+		Metrics:        []tools.CVSSMetric{},
+	}
+
+	intermediateData := IntermediateCVEDetailData{
+		DetailID:    dbRow.ID_2,
+		DetailCveID: dbRow.CveID_2,
+		Cwe:         dbRow.Cwe,
+		// ... Mirror ALL THE SAME field assignments as in the function above ...
+		PublishedDate:    dbRow.PublishedDate,
+		LastModifiedDate: dbRow.LastModifiedDate,
+		NvdDescription:   dbRow.NvdDescription,
+
+		CvssV2Vector:                dbRow.CvssV2Vector,
+		CvssV2BaseScore:             dbRow.CvssV2BaseScore,
+		CvssV2BaseSeverity:          dbRow.CvssV2BaseSeverity,
+		CvssV2ExploitabilityScore:   dbRow.CvssV2ExploitabilityScore,
+		CvssV2ImpactScore:           dbRow.CvssV2ImpactScore,
+		CvssV2AccessVector:          dbRow.CvssV2AccessVector,
+		CvssV2AccessComplexity:      dbRow.CvssV2AccessComplexity,
+		CvssV2Authentication:        dbRow.CvssV2Authentication,
+		CvssV2ConfidentialityImpact: dbRow.CvssV2ConfidentialityImpact,
+		CvssV2IntegrityImpact:       dbRow.CvssV2IntegrityImpact,
+		CvssV2AvailabilityImpact:    dbRow.CvssV2AvailabilityImpact,
+
+		CvssV30Vector:                dbRow.CvssV30Vector,
+		CvssV30BaseScore:             dbRow.CvssV30BaseScore,
+		CvssV30BaseSeverity:          dbRow.CvssV30BaseSeverity,
+		CvssV30ExploitabilityScore:   dbRow.CvssV30ExploitabilityScore,
+		CvssV30ImpactScore:           dbRow.CvssV30ImpactScore,
+		CvssV30AttackVector:          dbRow.CvssV30AttackVector,
+		CvssV30AttackComplexity:      dbRow.CvssV30AttackComplexity,
+		CvssV30PrivilegesRequired:    dbRow.CvssV30PrivilegesRequired,
+		CvssV30UserInteraction:       dbRow.CvssV30UserInteraction,
+		CvssV30Scope:                 dbRow.CvssV30Scope,
+		CvssV30ConfidentialityImpact: dbRow.CvssV30ConfidentialityImpact,
+		CvssV30IntegrityImpact:       dbRow.CvssV30IntegrityImpact,
+		CvssV30AvailabilityImpact:    dbRow.CvssV30AvailabilityImpact,
+
+		CvssV31Vector:                dbRow.CvssV31Vector,
+		CvssV31BaseScore:             dbRow.CvssV31BaseScore,
+		CvssV31BaseSeverity:          dbRow.CvssV31BaseSeverity,
+		CvssV31ExploitabilityScore:   dbRow.CvssV31ExploitabilityScore,
+		CvssV31ExploitCodeMaturity:   dbRow.CvssV31ExploitCodeMaturity,
+		CvssV31ImpactScore:           dbRow.CvssV31ImpactScore,
+		CvssV31AttackVector:          dbRow.CvssV31AttackVector,
+		CvssV31AttackComplexity:      dbRow.CvssV31AttackComplexity,
+		CvssV31PrivilegesRequired:    dbRow.CvssV31PrivilegesRequired,
+		CvssV31UserInteraction:       dbRow.CvssV31UserInteraction,
+		CvssV31Scope:                 dbRow.CvssV31Scope,
+		CvssV31ConfidentialityImpact: dbRow.CvssV31ConfidentialityImpact,
+		CvssV31IntegrityImpact:       dbRow.CvssV31IntegrityImpact,
+		CvssV31AvailabilityImpact:    dbRow.CvssV31AvailabilityImpact,
+
+		EpssScore:      dbRow.EpssScore,
+		EpssPercentile: dbRow.EpssPercentile,
+		RiskScore:      dbRow.RiskScore,
+		Likelihood:     dbRow.Likelihood,
+
+		EvaluatorComment:  dbRow.EvaluatorComment,
+		EvaluatorImpact:   dbRow.EvaluatorImpact,
+		EvaluatorSolution: dbRow.EvaluatorSolution,
+
+		NvdReferences:  dbRow.NvdReferences,
+		VendorComments: dbRow.VendorComments,
+	}
+
+	// Check if cve_details part exists (e.g. ID_2 is c.id from cve_details)
+	if intermediateData.DetailID.Valid {
+		populateVulnerabilityFromIntermediateDetails(&domVuln, &intermediateData)
+	} else {
+		domVuln.BaseSeverity = enums.SeverityTypeUnknown // Default if no CVE details
+	}
+
+	return domVuln
+}
+
+// IntermediateCVEDetailData holds the fields from the cve_details table
+// as they are selected by your sqlc queries.
+type IntermediateCVEDetailData struct {
+	DetailID         uuid.NullUUID
+	DetailCveID      sql.NullString
+	Cwe              sql.NullString
+	PublishedDate    sql.NullTime
+	LastModifiedDate sql.NullTime
+	NvdDescription   sql.NullString
+
+	CvssV2Vector                sql.NullString
+	CvssV2BaseScore             sql.NullString
+	CvssV2BaseSeverity          sql.NullString
+	CvssV2ExploitabilityScore   sql.NullString
+	CvssV2ImpactScore           sql.NullString
+	CvssV2AccessVector          sql.NullString
+	CvssV2AccessComplexity      sql.NullString
+	CvssV2Authentication        sql.NullString
+	CvssV2ConfidentialityImpact sql.NullString
+	CvssV2IntegrityImpact       sql.NullString
+	CvssV2AvailabilityImpact    sql.NullString
+
+	CvssV30Vector                sql.NullString
+	CvssV30BaseScore             sql.NullString
+	CvssV30BaseSeverity          sql.NullString
+	CvssV30ExploitabilityScore   sql.NullString
+	CvssV30ImpactScore           sql.NullString
+	CvssV30AttackVector          sql.NullString
+	CvssV30AttackComplexity      sql.NullString
+	CvssV30PrivilegesRequired    sql.NullString
+	CvssV30UserInteraction       sql.NullString
+	CvssV30Scope                 sql.NullString
+	CvssV30ConfidentialityImpact sql.NullString
+	CvssV30IntegrityImpact       sql.NullString
+	CvssV30AvailabilityImpact    sql.NullString
+
+	CvssV31Vector                sql.NullString
+	CvssV31BaseScore             apd.NullDecimal
+	CvssV31BaseSeverity          sql.NullString
+	CvssV31ExploitabilityScore   apd.NullDecimal
+	CvssV31ExploitCodeMaturity   sql.NullString
+	CvssV31ImpactScore           apd.NullDecimal
+	CvssV31AttackVector          sql.NullString
+	CvssV31AttackComplexity      sql.NullString
+	CvssV31PrivilegesRequired    sql.NullString
+	CvssV31UserInteraction       sql.NullString
+	CvssV31Scope                 sql.NullString
+	CvssV31ConfidentialityImpact sql.NullString
+	CvssV31IntegrityImpact       sql.NullString
+	CvssV31AvailabilityImpact    sql.NullString
+
+	EpssScore      sql.NullString
+	EpssPercentile sql.NullString
+	RiskScore      apd.NullDecimal
+	Likelihood     sql.NullString
+
+	EvaluatorComment  sql.NullString
+	EvaluatorImpact   sql.NullString
+	EvaluatorSolution sql.NullString
+
+	NvdReferences  pqtype.NullRawMessage
+	VendorComments pqtype.NullRawMessage
+}
+
+// populateVulnerabilityFromIntermediateDetails maps data from IntermediateCVEDetailData
+// to the tools.Vulnerability struct, focusing on Metrics and derived top-level fields.
+func populateVulnerabilityFromIntermediateDetails(domVuln *tools.Vulnerability, cveData *IntermediateCVEDetailData) {
+	if cveData == nil {
+		return // Nothing to map from cve_details
+	}
+
+	// If domVuln.Description is from v.description, and cveData.NvdDescription is preferred:
+	domVuln.Description = nullStringToString(cveData.NvdDescription)
+	domVuln.Published = nullTimeToTime(cveData.PublishedDate)
+	domVuln.LastUpdated = nullTimeToTime(cveData.LastModifiedDate)
+
+	if cweType, ok := enums.ParseWeaknessFromString(cveData.Cwe.String); ok {
+		domVuln.Type = cweType
+	} else if cveData.Cwe.Valid {
+		slog.Warn("Unknown CWE/WeaknessType string from DB", "cwe", cveData.Cwe.String, "cve_id", domVuln.CveID)
+		domVuln.Type = enums.WeaknessNoInfo
+	}
+
+	if refs, err := unmarshalNvdReferences(cveData.NvdReferences); err == nil {
+		domVuln.References = refs
+	} else if cveData.NvdReferences.Valid {
+		slog.Warn("Failed to unmarshal NvdReferences", "cve_id", domVuln.CveID, "error", err)
+	}
+	if vc, err := unmarshalNvdVendorComments(cveData.VendorComments); err == nil {
+		domVuln.VendorComments = vc
+	} else if cveData.VendorComments.Valid {
+		slog.Warn("Failed to unmarshal VendorComments", "cve_id", domVuln.CveID, "error", err)
+	}
+
+	var metrics []tools.CVSSMetric
+
+	// CVSS v2.0
+	if cveData.CvssV2BaseScore.Valid || cveData.CvssV2Vector.Valid {
+		m2 := tools.CVSSMetric{Version: enums.CVSSv20}
+		if s, err := parseNullStringAsFloat64(cveData.CvssV2BaseScore); err == nil {
+			m2.BaseScore = s
+		}
+		m2.Severity = enums.ParseSeverityType(cveData.CvssV2BaseSeverity.String, enums.SeverityTypeUnknown)
+		if s, err := parseNullStringAsFloat64(cveData.CvssV2ExploitabilityScore); err == nil {
+			m2.ExploitabilityScore = s
+		}
+		if s, err := parseNullStringAsFloat64(cveData.CvssV2ImpactScore); err == nil {
+			m2.ImpactScore = s
+		}
+		m2.Access = enums.ParseAccessType(cveData.CvssV2AccessVector.String, enums.AccessTypeUnknown)
+		m2.Complexity = enums.ParseComplexityType(cveData.CvssV2AccessComplexity.String, enums.ComplexityTypeUnknown) // Ensure ComplexityType has LOW/MEDIUM/HIGH
+		m2.IntegrityImpact = enums.ParseImpactType(cveData.CvssV2IntegrityImpact.String, enums.ImpactTypeUnknown)
+		m2.AvailabilityImpact = enums.ParseImpactType(cveData.CvssV2AvailabilityImpact.String, enums.ImpactTypeUnknown)
+		metrics = append(metrics, m2)
+	}
+
+	// CVSS v3.0
+	if cveData.CvssV30BaseScore.Valid || cveData.CvssV30Vector.Valid {
+		m30 := tools.CVSSMetric{Version: enums.CVSSv30}
+		if s, err := parseNullStringAsFloat64(cveData.CvssV30BaseScore); err == nil {
+			m30.BaseScore = s
+		}
+		m30.Severity = enums.ParseSeverityType(cveData.CvssV30BaseSeverity.String, enums.SeverityTypeUnknown)
+		if s, err := parseNullStringAsFloat64(cveData.CvssV30ExploitabilityScore); err == nil {
+			m30.ExploitabilityScore = s
+		}
+		if s, err := parseNullStringAsFloat64(cveData.CvssV30ImpactScore); err == nil {
+			m30.ImpactScore = s
+		}
+		m30.Access = enums.ParseAccessType(cveData.CvssV30AttackVector.String, enums.AccessTypeUnknown)
+		m30.Complexity = enums.ParseComplexityType(cveData.CvssV30AttackComplexity.String, enums.ComplexityTypeUnknown)
+		m30.PrivilegesRequired = enums.ParsePrivilegesRequiredType(cveData.CvssV30PrivilegesRequired.String, enums.PrivilegesRequiredUnknown)
+		m30.IntegrityImpact = enums.ParseImpactType(cveData.CvssV30IntegrityImpact.String, enums.ImpactTypeUnknown)
+		m30.AvailabilityImpact = enums.ParseImpactType(cveData.CvssV30AvailabilityImpact.String, enums.ImpactTypeUnknown)
+		metrics = append(metrics, m30)
+	}
+
+	// CVSS v3.1
+	if cveData.CvssV31BaseScore.Valid {
+		m31 := tools.CVSSMetric{Version: enums.CVSSv31}
+		if s, err := nullDecimalToFloat64(cveData.CvssV31BaseScore); err == nil {
+			m31.BaseScore = s
+		}
+		m31.Severity = enums.ParseSeverityType(cveData.CvssV31BaseSeverity.String, enums.SeverityTypeUnknown)
+		if s, err := nullDecimalToFloat64(cveData.CvssV31ExploitabilityScore); err == nil {
+			m31.ExploitabilityScore = s
+		}
+		m31.Exploitability = enums.ParseExploitabilityType(cveData.CvssV31ExploitCodeMaturity.String, enums.ExploitabilityTypeUnknown)
+		if s, err := nullDecimalToFloat64(cveData.CvssV31ImpactScore); err == nil {
+			m31.ImpactScore = s
+		}
+		m31.Access = enums.ParseAccessType(cveData.CvssV31AttackVector.String, enums.AccessTypeUnknown)
+		m31.Complexity = enums.ParseComplexityType(cveData.CvssV31AttackComplexity.String, enums.ComplexityTypeUnknown)
+		m31.PrivilegesRequired = enums.ParsePrivilegesRequiredType(cveData.CvssV31PrivilegesRequired.String, enums.PrivilegesRequiredUnknown)
+		m31.IntegrityImpact = enums.ParseImpactType(cveData.CvssV31IntegrityImpact.String, enums.ImpactTypeUnknown)
+		m31.AvailabilityImpact = enums.ParseImpactType(cveData.CvssV31AvailabilityImpact.String, enums.ImpactTypeUnknown)
+		metrics = append(metrics, m31)
+	}
+	domVuln.Metrics = metrics
+
+	// Populate Top-Level Convenience Fields from the "best" Metric
+	var primaryMetric *tools.CVSSMetric
+	// Prioritize v3.1 -> v3.0 -> v2.0
+	for i := range domVuln.Metrics {
+		mPtr := &domVuln.Metrics[i] // Get pointer for direct comparison
+		if mPtr.Version == enums.CVSSv31 {
+			primaryMetric = mPtr
+			break
+		}
+	}
+	if primaryMetric == nil {
+		for i := range domVuln.Metrics {
+			mPtr := &domVuln.Metrics[i]
+			if mPtr.Version == enums.CVSSv30 {
+				primaryMetric = mPtr
+				break
+			}
+		}
+	}
+	if primaryMetric == nil && len(domVuln.Metrics) > 0 {
+		primaryMetric = &domVuln.Metrics[0] // Fallback to the first one if no v3.x found
+	}
+
+	if primaryMetric != nil {
+		domVuln.BaseCVSSScore = primaryMetric.BaseScore
+		domVuln.ImpactScore = primaryMetric.ImpactScore
+		domVuln.BaseSeverity = primaryMetric.Severity
+		domVuln.Access = primaryMetric.Access
+		domVuln.Complexity = primaryMetric.Complexity
+		domVuln.PrivilegesRequired = primaryMetric.PrivilegesRequired
+		domVuln.IntegrityImpact = primaryMetric.IntegrityImpact
+		domVuln.AvailabilityImpact = primaryMetric.AvailabilityImpact
+		domVuln.Exploit.Score = primaryMetric.ExploitabilityScore
+		domVuln.Exploit.Exploitability = primaryMetric.Exploitability
+	} else {
+		// If no metrics at all, top-level fields remain zero/default.
+		// Optionally, log this situation.
+		if domVuln.CveID != "" { // Only log if it's a known CVE that surprisingly has no metrics
+			slog.Info("No CVSS metrics found to populate top-level fields", "cve_id", domVuln.CveID)
+		}
+	}
+
+	// Map other direct fields
+	domVuln.Likelihood = enums.ParseLikelyhoodType(cveData.Likelihood.String, enums.LikelyhoodTypeUnknown)
+	if score, err := nullDecimalToFloat64(cveData.RiskScore); err == nil && cveData.RiskScore.Valid {
+		domVuln.RiskScore = score
+	} else if cveData.RiskScore.Valid {
+		slog.Warn("Error converting RiskScore from apd.Decimal", "cve_id", domVuln.CveID, "error", err)
+	}
+
+	if score, err := parseNullStringAsFloat64(cveData.EpssScore); err == nil && cveData.EpssScore.Valid {
+		domVuln.EPSSScore = score
+	} else if cveData.EpssScore.Valid {
+		slog.Warn("Error parsing EpssScore", "cve_id", domVuln.CveID, "value", cveData.EpssScore.String, "error", err)
+	}
+	if score, err := parseNullStringAsFloat64(cveData.EpssPercentile); err == nil && cveData.EpssPercentile.Valid {
+		domVuln.EPSSPercentile = score
+	} else if cveData.EpssPercentile.Valid {
+		slog.Warn("Error parsing EpssPercentile", "cve_id", domVuln.CveID, "value", cveData.EpssPercentile.String, "error", err)
+	}
+}
+
+func buildSeverityWhereClause(severityFilters []string, queryParams []any) (string, []any) {
+	severityWhereClause := ""
+	if len(severityFilters) > 0 {
+		placeholders := make([]string, len(severityFilters))
+		for i, severity := range severityFilters {
+			placeholders[i] = fmt.Sprintf("$%d", len(queryParams)+1)
+			queryParams = append(queryParams, severity)
+		}
+		severityWhereClause = fmt.Sprintf("AND sv.severity ILIKE ANY(array[%s])", strings.Join(placeholders, ","))
+	}
+
+	return severityWhereClause, queryParams
 }
